Centered Reset Button and Made it Work
Got the reset button actually working by adding another property to the classes that records what the numbers for the stats were originally, then sets the stats to that when the reset button is pressed. I thought this would have happened automatically because it shouldn't have been changing the stats of the classes object template, but rather the player list which was just a copy. JS must work in a different way I guess.

Also added CSS to the reset button and centered it.

I decided trying to make a "class" class for the different "character" classes was more effort than it was worth. It works fine the way it is now.